fix(committersStore): change directory structure
