+++ static/app/stores/Commiters/committersContext.tmp.tsx
@@ -1,0 +1,7 @@
+import {createContext} from 'react';
+
+import {CommittersAction, CommittersState} from './committersReducer';
+
+export const CommittersContext = createContext<
+  [CommittersState, React.Dispatch<CommittersAction>] | null
+>(null);
